Code refactoring + new github copilot instructions
