+++ .github/copilot-instructions.md
@@ -1,0 +1,35 @@
+# GitHub Copilot Instructions
+
+## General Guidelines
+
+- Always provide clean, efficient, and well-documented code solutions
+- Focus on best practices and modern coding standards
+- Ensure code is secure, maintainable, and follows industry standards
+
+## Restrictions
+
+- Do not provide any commands to execute in the terminal or command line
+- Do not suggest destructive operations or changes to system files
+- Do not generate code that could potentially harm systems or data
+- Do not provide solutions that violate licensing terms or copyrights
+
+## Preferred Coding Style
+
+- Use clear variable/function naming that reflects purpose
+- Include appropriate comments explaining complex logic
+- Favor readability over clever/complex solutions
+- Follow the project's existing code style when applicable
+- Organize code into logical sections with appropriate separation
+
+## Documentation
+
+- Include relevant documentation references when appropriate
+- Explain key design decisions in comments
+- Document any assumptions made in the solution
+- Provide usage examples for complex functions/classes
+
+## Error Handling
+
+- Always include proper error handling in generated code
+- Avoid suppressing errors without explanation
+- Provide guidance on potential error scenarios and how to handle them
