+++ zfs-assistant/REORGANIZATION.md
@@ -1,0 +1,77 @@
+# ZFS Assistant Codebase Reorganization - Summary
+
+## Overview
+
+The ZFS Assistant codebase has been reorganized from a flat structure into a modular, package-based architecture. This reorganization improves maintainability, readability, and separation of concerns.
+
+## Changes Made
+
+### 1. Created Logical Package Structure
+
+| Package | Purpose |
+|---------|---------|
+| `core/` | Core ZFS operations (dataset management, snapshots) |
+| `backup/` | Backup and restore functionality (send/receive) |
+| `system/` | System integration (pacman hooks, systemd timers) and maintenance |
+| `ui/` | User interface components |
+| `utils/` | Common utilities, models, logging |
+
+### 2. UI Reorganization
+
+The UI components have been further organized into:
+
+- `ui/windows/` - Main application windows
+- `ui/dialogs/` - Modal dialog windows
+- `ui/settings/` - Settings-related UI components
+- `ui/components/` - Reusable UI components
+
+### 3. Detailed Changes
+
+1. **Core Module:**
+   - Moved `zfs_core.py` to `core/zfs_core.py`
+   - Updated imports to use relative paths
+
+2. **Backup Module:**
+   - Moved `zfs_backup.py` to `backup/zfs_backup.py`
+   - Updated imports to use relative paths
+
+3. **System Module:**
+   - Moved `system_integration.py` to `system/system_integration.py`
+   - Moved `system_maintenance.py` to `system/system_maintenance.py`
+   - Updated imports to use relative paths
+
+4. **Utils Module:**
+   - Moved `common.py` to `utils/common.py`
+   - Moved `logger.py` to `utils/logger.py`
+   - Moved `models.py` to `utils/models.py`
+   - Moved `privilege_manager.py` to `utils/privilege_manager.py`
+
+5. **UI Module:**
+   - Previously reorganized the UI files with a similar structure
+   - Updated all imports to reflect new organization
+
+6. **Import Updates:**
+   - Updated imports in all files to reflect the new structure
+   - Modified `zfs_assistant.py` to use the new import paths
+   - Updated `application.py` and `__main__.py` with new import paths
+
+7. **Package Files:**
+   - Created `__init__.py` files in each directory to make them proper Python packages
+   - Configured exports in `__init__.py` files to simplify imports
+
+8. **Setup.py:**
+   - Updated to use `find_packages()` to automatically find the new package structure
+
+## Benefits of the New Structure
+
+1. **Separation of Concerns:** Each module has a clear responsibility
+2. **Better Maintainability:** Smaller, focused files are easier to understand and modify
+3. **Improved Organization:** Logical grouping of related functionality
+4. **Easier Navigation:** Clear folder structure makes finding code easier
+5. **Proper Python Packaging:** Using standard Python package conventions
+
+## Next Steps
+
+1. **Testing:** Thoroughly test the application to ensure it works with the new structure
+2. **Documentation:** Update any remaining documentation to reflect the new structure
+3. **Ongoing Improvements:** Continue to refactor and improve code within each module
