+++ zfs-assistant/src/backup/__init__.py
@@ -1,0 +1,14 @@
+#!/usr/bin/env python3
+# ZFS Assistant - Backup Module
+# Author: GitHub Copilot
+
+"""
+ZFS Backup Module
+
+Contains functionality for backup, restore, and transfer operations
+between ZFS pools and datasets.
+"""
+
+from .zfs_backup import ZFSBackup
+
+__all__ = ['ZFSBackup']
