+++ zfs-assistant/src/core/__init__.py
@@ -1,0 +1,14 @@
+#!/usr/bin/env python3
+# ZFS Assistant - Core Module
+# Author: GitHub Copilot
+
+"""
+ZFS Core Module
+
+Contains core functionality for ZFS operations like dataset management,
+snapshot creation, and rollback operations.
+"""
+
+from .zfs_core import ZFSCore
+
+__all__ = ['ZFSCore']
