+++ zfs-assistant/src/core/zfs_core.py
@@ -5,12 +5,12 @@
 import subprocess
 import datetime
 from typing import List, Optional, Tuple
-from .models import ZFSSnapshot
-from .logger import (
+from ..utils.models import ZFSSnapshot
+from ..utils.logger import (
     OperationType, LogLevel, get_logger,
     log_info, log_error, log_success, log_warning
 )
-from .common import get_timestamp
+from ..utils.common import get_timestamp
 
 
 class ZFSCore:
