+++ zfs-assistant/src/system/__init__.py
@@ -1,0 +1,15 @@
+#!/usr/bin/env python3
+# ZFS Assistant - System Module
+# Author: GitHub Copilot
+
+"""
+ZFS System Module
+
+Contains functionality for system integration, maintenance, updates,
+and package management related to ZFS.
+"""
+
+from .system_integration import SystemIntegration
+from .system_maintenance import SystemMaintenance
+
+__all__ = ['SystemIntegration', 'SystemMaintenance']
