+++ zfs-assistant/src/ui/components/__init__.py
@@ -1,0 +1,19 @@
+#!/usr/bin/env python3
+# ZFS Assistant - UI Components Package
+# Author: GitHub Copilot
+
+"""
+UI Components Package
+
+Contains reusable UI components and widgets that can be used across multiple windows and dialogs.
+"""
+
+from .shared import (
+    FrameWithContent, ButtonBox, LabeledEntry, 
+    LabeledSwitch, InfoLabel, ResponsiveGrid
+)
+
+__all__ = [
+    'FrameWithContent', 'ButtonBox', 'LabeledEntry', 
+    'LabeledSwitch', 'InfoLabel', 'ResponsiveGrid'
+]
