+++ zfs-assistant/src/ui/dialogs/__init__.py
@@ -1,0 +1,13 @@
+#!/usr/bin/env python3
+# ZFS Assistant - UI Dialogs Package
+# Author: GitHub Copilot
+
+"""
+UI Dialogs Package
+
+Contains modal dialog windows and popup components.
+"""
+
+from .create_snapshot_dialog import CreateSnapshotDialog
+
+__all__ = ['CreateSnapshotDialog']
