+++ zfs-assistant/src/ui/windows/__init__.py
@@ -1,0 +1,13 @@
+#!/usr/bin/env python3
+# ZFS Assistant - UI Windows Package
+# Author: GitHub Copilot
+
+"""
+UI Windows Package
+
+Contains main application windows and top-level UI containers.
+"""
+
+from .main_window import MainWindow
+
+__all__ = ['MainWindow']
