+++ zfs-assistant/src/utils/__init__.py
@@ -1,0 +1,30 @@
+#!/usr/bin/env python3
+# ZFS Assistant - Utils Module
+# Author: GitHub Copilot
+
+"""
+ZFS Utils Module
+
+Contains utility functions, models, logging, and common functionality
+used across other ZFS Assistant modules.
+"""
+
+from .common import (
+    CONFIG_DIR, CONFIG_FILE, LOG_FILE, PACMAN_HOOK_PATH, 
+    SYSTEMD_SCRIPT_PATH, PACMAN_SCRIPT_PATH,
+    DEFAULT_CONFIG, get_timestamp
+)
+from .logger import (
+    ZFSLogger, OperationType, LogLevel, get_logger,
+    log_info, log_error, log_success, log_warning
+)
+from .models import ZFSSnapshot
+from .privilege_manager import PrivilegeManager
+
+__all__ = [
+    'CONFIG_DIR', 'CONFIG_FILE', 'LOG_FILE', 'PACMAN_HOOK_PATH', 
+    'SYSTEMD_SCRIPT_PATH', 'PACMAN_SCRIPT_PATH', 'DEFAULT_CONFIG', 
+    'get_timestamp', 'ZFSLogger', 'OperationType', 'LogLevel', 
+    'get_logger', 'log_info', 'log_error', 'log_success', 
+    'log_warning', 'ZFSSnapshot', 'PrivilegeManager'
+]
