+++ zfs-assistant/src/utils/privilege_manager.py
